refactor(teacher): TeacherStats sous 300 lignes — éclatement complet (H11)
Vue 698 → orchestrateur 151 l. Logique extraite dans useTeacherStats.js
(mapping dashboard→compteurs via helper partagé loadStats/refresh, cache +
rafraîchissement en arrière-plan). Template scindé en 5 composants :
TeacherStatsGlobalCards, TeacherStatsPerMatiere, TeacherStatsActivity,
TeacherStatsParClasse et TeacherStatsQuickActions. CSS déplacé verbatim
(chrome widget + bg-*-100 dupliqués par composant scopé, @media éclatés).
Parité : mêmes appels services/cache, mêmes classes/textes. La duplication
du mapping (loadStats/refreshInBackground) est convergée dans un helper
mapDashboardToStats — sortie strictement identique.
Tests : composable (3) + 5 montages.
